refactor: Frames를 Player로 변경, 기존의 Player 제거
